+++ demo/src/app/components/+vote/vote.component.ts
@@ -2,6 +2,7 @@ import { Component } from '@angular/core';
 import { LiveDemoCodeExample } from 'app/core/live-demo/live-demo.component';
 import { DemoVoteBasicComponent } from './basic/vote-basic.component';
 import { DemoVoteWeakComponent } from './weak/vote-weak.component';
+import { DemoVoteIconComponent } from './icon/vote-icon.component';
 
 @Component({
     selector: 'demo-vote',
@@ -22,11 +23,10 @@ export class DemoVoteComponent {
             default: 'horizontal'
         },
         {
-            property: 'thySizs',
-            description:
-                'thyLayout="vertical"支持"sm"和"md",thyLayout="horizontal"支持"sm",只限制高度，宽度根据数字自适应',
-            type: 'String',
-            default: 'sm'
+            property: 'thySize',
+            description: 'thyLayout="vertical"支持"sm"和"default"',
+            type: 'string',
+            default: 'default'
         },
         {
             property: 'thyHasVoted',
@@ -39,6 +39,18 @@ export class DemoVoteComponent {
             description: '赞同的数量',
             type: 'number | string',
             default: ''
+        },
+        {
+            property: 'thyIcon',
+            description: '图标',
+            type: 'string',
+            default: 'thumb-up'
+        },
+        {
+            property: 'thyRound',
+            description: '是否是偏圆型',
+            type: 'boolean',
+            default: 'false'
         }
     ];
     liveDemos: LiveDemoCodeExample[] = [
@@ -71,6 +83,22 @@ export class DemoVoteComponent {
                     type: 'ts',
                     name: 'vote-weak.component.ts',
                     content: require('!!raw-loader!./weak/vote-weak.component.ts')
+                }
+            ]
+        },
+        {
+            title: 'Vote Icon',
+            component: DemoVoteIconComponent,
+            codeExamples: [
+                {
+                    type: 'html',
+                    name: 'vote-icon.component.html',
+                    content: require('!!raw-loader!./icon/vote-icon.component.html')
+                },
+                {
+                    type: 'ts',
+                    name: 'vote-icon.component.ts',
+                    content: require('!!raw-loader!./icon/vote-icon.component.ts')
                 }
             ]
         }