+++ src/mention/interfaces.ts
@@ -1,4 +1,5 @@
 import { TemplateRef } from '@angular/core';
+import { Observable } from 'rxjs';
 
 export interface MentionDefaultDataItem {
     name?: string;
@@ -9,6 +10,13 @@ export interface Mention<T = any> {
     data: T[];
     limit?: number;
     insertTransform?: (item: T) => string;
-    search?: (term: string, items: T[]) => T[];
+    search?: (term: string, items?: T[]) => T[] | Observable<T[]>;
     displayTemplateRef?: TemplateRef<T>;
+    emptyText?: string;
+    popoverClass?: string;
+}
+
+export interface MentionSuggestionSelectEvent<T = any> {
+    event: Event;
+    item: T;
 }