+++ src/popover/popover.service.ts
@@ -4,7 +4,8 @@ import {
     OverlayConfig,
     OverlayRef,
     PositionStrategy,
-    ScrollDispatcher
+    ScrollDispatcher,
+    OverlayContainer
 } from '@angular/cdk/overlay';
 import {
     TemplateRef,
@@ -27,6 +28,13 @@ import { getFlexiblePositions, ThyUpperOverlayService, ThyUpperOverlayRef } from
 import { takeUntil } from 'rxjs/operators';
 import { helpers } from '../util';
 import { popoverUpperOverlayOptions } from './popover.options';
+import { ViewportRuler } from '@angular/cdk/scrolling';
+import { DOCUMENT } from '@angular/platform-browser';
+import { Platform } from '@angular/cdk/platform';
+import {
+    FlexibleConnectedPositionStrategy,
+    FlexibleConnectedPositionStrategyOrigin
+} from '../core/overlay/position/flexible-connected-position-strategy';
 
 @Injectable({
     providedIn: 'root'
@@ -44,22 +52,27 @@ export class ThyPopover extends ThyUpperOverlayService<ThyPopoverConfig, ThyPopo
     >();
 
     private buildPositionStrategy<TData>(config: ThyPopoverConfig<TData>): PositionStrategy {
-        if (config.position) {
-            const positionStrategy = this.overlay.position().global();
-            return positionStrategy;
-        } else {
-            const positionStrategy = this.overlay.position().flexibleConnectedTo(coerceElement(config.origin));
-            const positions = getFlexiblePositions(config.placement, config.offset, 'thy-popover');
-            positionStrategy.withPositions(positions);
-            positionStrategy.positionChanges.pipe(takeUntil(this.ngUnsubscribe$)).subscribe(change => {
-                if (change.scrollableViewProperties.isOverlayClipped) {
-                    // After position changes occur and the overlay is clipped by
-                    // a parent scrollable then close the tooltip.
-                    this.ngZone.run(() => this.close());
-                }
-            });
-            return positionStrategy;
-        }
+        const origin: FlexibleConnectedPositionStrategyOrigin = config.originPosition
+            ? config.originPosition
+            : config.origin;
+        // const positionStrategy = this.overlay.position().flexibleConnectedTo(origin);
+        const positionStrategy = new FlexibleConnectedPositionStrategy(
+            origin,
+            this._viewportRuler,
+            this._document,
+            this._platform,
+            this._overlayContainer
+        );
+        const positions = getFlexiblePositions(config.placement, config.offset, 'thy-popover');
+        positionStrategy.withPositions(positions);
+        positionStrategy.positionChanges.pipe(takeUntil(this.ngUnsubscribe$)).subscribe(change => {
+            if (change.scrollableViewProperties.isOverlayClipped) {
+                // After position changes occur and the overlay is clipped by
+                // a parent scrollable then close the tooltip.
+                this.ngZone.run(() => this.close());
+            }
+        });
+        return positionStrategy;
     }
 
     private buildOverlayPanelClasses(config: ThyPopoverConfig) {
@@ -141,7 +154,11 @@ export class ThyPopover extends ThyUpperOverlayService<ThyPopoverConfig, ThyPopo
         injector: Injector,
         @Inject(THY_POPOVER_DEFAULT_CONFIG) defaultConfig: ThyPopoverConfig,
         private scrollDispatcher: ScrollDispatcher,
-        private ngZone: NgZone
+        private ngZone: NgZone,
+        private _viewportRuler: ViewportRuler,
+        @Inject(DOCUMENT) private _document: any,
+        private _platform: Platform,
+        private _overlayContainer: OverlayContainer
     ) {
         super(popoverUpperOverlayOptions, overlay, injector, defaultConfig);
     }
@@ -176,7 +193,7 @@ export class ThyPopover extends ThyUpperOverlayService<ThyPopoverConfig, ThyPopo
             return;
         }
 
-        const popoverRef = this.openUpperOverlay(componentOrTemplateRef, config);
+        const popoverRef = this.openUpperOverlay(componentOrTemplateRef, config) as ThyPopoverRef<T>;
         config = popoverRef.containerInstance.config;
         popoverRef.afterClosed().subscribe(() => {
             this.originElementRemoveActiveClass(config);
@@ -184,9 +201,6 @@ export class ThyPopover extends ThyUpperOverlayService<ThyPopoverConfig, ThyPopo
         });
 
         this.originElementAddActiveClass(config);
-        if (config.position) {
-            popoverRef.updatePosition(config.position);
-        }
         this.originInstancesMap.set(originElement, {
             config,
             popoverRef